requirements.txt
aiogram==3.2.0
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
python-dotenv==1.0.0
gtts==2.4.0
requests==2.31.0
schedule==1.2.0
asyncio-mqtt==0.11.1
aiofiles==23.2.1
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.25.2
pandas==2.1.4
psycopg2-binary==2.9.9
uvloop==0.19.0